[Feature]: Add Validation for wrapper functions and Github templates

+++ src/Methods/commentsMethods/deleteSingleComment.js
@@ -1,15 +1,18 @@
 const rp = require('request-promise');
 
-const deleteSingleComment = async (params, commentId) => {
+const deleteSingleComment = async (token, commentId, userId) => {
+  if (!userId) {
+    return `User ID must not be empty!`;
+  }
   var options = {
     uri: `https://comment.microapi.dev/v1/comments/${commentId}`,
     headers: {
       'User-Agent': 'Request-Promise',
-      Authorization: `Bearer ${params.appToken}`,
+      Authorization: `Bearer ${token}`,
     },
     method: 'DELETE',
     body: {
-      ownerId: params.ownerId,
+      ownerId: userId,
     },
     json: true, // Automatically parses the JSON string in the response
   };
